pug-or-ugh_v1/requirements.txt
coverage==4.5.3
Django==2.1.5
djangorestframework==3.9.2
pytz==2018.9
sqlparse==0.3.0
whitenoise==4.1.2